Add x-set-delay command
This command adds a configurable delay (in milliseconds) between the
mouse down event and the mouse up event. This fixes issues with certain
applications that require a delay between the two events for them to
register as a click.

Note that this command is non-standard (i.e. not implemented in the
original keynav).

Fixes #15